Only run mri_convert once per input file.
If fresh is split to run the longitiduinal and recon-all -all steps
separtely it could end up running mri_convert more than once on the same
input files and over-writing the output files.

This changes introduces a conditional check to ensure if the putput file
already exists the mri_convert step is not run.